+++ resources/install.rb
@@ -1,3 +1,5 @@
+provides :nginx_install
+
 property :ohai_plugin_enabled, [true, false],
          description: 'Whether or not ohai_plugin is enabled.',
          equal_to: [true, false],
@@ -179,8 +181,7 @@ action :install do
         baseurl      repo_url
         gpgkey       repo_signing_key
       end
-      execute 'disable-nginx-module' do
-        command 'dnf -y module disable nginx'
+      execute 'dnf -qy module disable nginx' do
         only_if { node['platform_version'].to_i >= 8 && platform_family?('rhel') || platform_family?('fedora') }
         not_if 'dnf module list nginx | grep -q "^nginx.*\[x\]"'
       end